migration to vite from create-react-app
