+++ frontend/package.json
@@ -34,6 +34,7 @@
     "yup": "^1.3.2"
   },
   "scripts": {
+    "dev": "vite",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
@@ -59,8 +60,10 @@
     ]
   },
   "devDependencies": {
+    "@vitejs/plugin-react-swc": "^3.4.0",
     "eslint": "^8.52.0",
     "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0"
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "vite": "^4.5.0"
   }
 }
