+++ frontend/src/components/HistoryCard.jsx
