+++ frontend/src/components/LandingPage/Hero.jsx
