+++ frontend/src/components/LandingPage/footer_segment/BoostBox.jsx
