+++ frontend/src/components/LandingPage/footer_segment/FooterLink.jsx
