+++ frontend/src/components/LandingPage/footer_segment/FooterSegment.jsx
