+++ frontend/src/components/LandingPage/form/SingleOutput.jsx
